+++ webpack/styleguide.config.js
@@ -1,40 +1,44 @@
 /* eslint-disable import/no-commonjs */
 /* eslint-env node */
 
-const ExtractTextPlugin = require('extract-text-webpack-plugin')
+const path = require('path')
 
 module.exports = {
-    title: 'Progressive Web SDK',
-    components: './app/components/**/*.jsx',
+    title: 'Progressive Web Scaffold',
+    components: '../app/components/**/*.jsx',
     serverHost: '0.0.0.0',
     serverPort: 4000,
     skipComponentsWithoutExample: true,
     updateWebpackConfig(webpackConfig) {
-        // Plugins
-        webpackConfig.plugins.push(
-            new ExtractTextPlugin('css/[name].css')
-        )
+        // Supply our own renderer for styleguide
+        webpackConfig.resolve.alias['rsg-components/Layout/Renderer'] = path.join(__dirname, '../styleguide/renderer')
 
         // Loaders
         webpackConfig.module.loaders.push(
             {
                 test: /\.jsx?$/,
                 exclude: /node_modules/,
-                include: __dirname,
                 loader: 'babel',
                 query: {
                     presets: ['es2015', 'react']
                 },
-                cacheDirectory: `${__dirname}/tmp`
             },
+            // Loader for styleguide & project styles
             {
                 test: /\.scss$/,
-                loader: ExtractTextPlugin.extract(['css', 'sass']),
+                loader: 'style!css!sass',
                 include: [
-                    /progressive-web-sdk/,
-                    /app/
+                    /node_modules\/progressive-web-sdk/,
+                    /styleguide/
                 ]
             },
+            // Loader for plain CSS
+            {
+                test: /\.css$/,
+                loader: 'style!css?modules&importLoaders=1',
+                include: /styleguide/,
+            },
+            // Loader for SVGs
             {
                 test: /\.svg$/,
                 loader: 'text',