+++ CHANGELOG.md
@@ -1,5 +1,13 @@
+## To be released
+- Add PostCSS (Autoprefixer) and CSS minification
+
+## 0.0.4 (August 31)
+- Add configuration for building in specific cache manifest files into the main
+  and loader build files
+- Move Styleguidist's webpack settings to the `/webpack` directory 
+
 ## 0.0.3 (August 26)
- - use SDK version 0.1.1
+- use SDK version 0.1.1
 
 ## 0.0.2
 - use SDK version 0.0.11